# docs/api/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Widget Demo Dynamic Service",
        "description": "The ServiceWidgetDemo is a KBase SDK-compatible dynamic service for demonstrating\nthe ability to provide widgets in a dynamic service.\n",
        "termsOfService": "https://www.kbase.us/about/terms-and-conditions-v2/",
        "contact": {
            "name": "KBase, Lawrence Berkeley National Laboratory, DOE",
            "url": "https://www.kbase.us/",
            "email": "engage@kbase.us"
        },
        "license": {
            "name": "The MIT License",
            "url": "https://github.com/kbase/kb_sdk/blob/develop/LICENSE.md"
        },
        "version": "0.1.0"
    },
    "paths": {
...
        "/rcsb-annotations/{ref}": {
            "get": {
                "tags": [
                    "misc"
                ],
                "summary": "Get Rcsb Annotations",
                "description": "Get RCSB Annotations from a Genome",
                "operationId": "get_rcsb_annotations_rcsb_annotations__ref__get",
                "parameters": [
                    {
                        "name": "ref",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Ref"
                        }
                    },
                    {
                        "name": "authorization",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string",
                                    "minLength": 32,
                                    "maxLength": 32
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "KBase auth token",
                            "title": "Authorization"
                        },
                        "description": "KBase auth token"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "title": "Response Get Rcsb Annotations Rcsb Annotations  Ref  Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    },
...
}